Restructure app UI and reorganize project files.
