+++ MySaver/AppShell.xaml
@@ -5,20 +5,22 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:local="clr-namespace:MySaver"
     xmlns:views="clr-namespace:MySaver.Views"
-    Shell.FlyoutBehavior="Disabled">
-
-
+    Shell.FlyoutBehavior="Flyout">
 
     <TabBar>
-        <ShellContent Title="Home"
-                      ContentTemplate="{DataTemplate local:MainPage}"
-                      Route="MainPage"/>
-
-        <ShellContent Title="About"
-                      ContentTemplate="{DataTemplate local:AboutPage}"
-                      Route="AboutPage"/>
-        <ShellContent Title="User"
-                      ContentTemplate="{DataTemplate views:UserSummaryPage}"/>
+        <ShellContent ContentTemplate="{DataTemplate views:LoginPage}"
+                      Route="LoginPage"/>
     </TabBar>
+
+    <FlyoutItem Title="About">
+        <ShellContent ContentTemplate="{DataTemplate views:AboutPage}"
+                      Route="AboutPage"/>
+    </FlyoutItem>
+    <FlyoutItem Title="User">
+        <ShellContent ContentTemplate="{DataTemplate views:UserSummaryPage}"
+                      Route="UserSummaryPage"/>
+    </FlyoutItem>
+    <MenuItem Text="Logout"
+              Clicked="OnSignOutClicked"/>
 
 </Shell>
