--- MySaver/MainPage.xaml.cs
@@ -1,16 +1,0 @@
-﻿namespace MySaver;
-
-public partial class MainPage : ContentPage
-{
-
-	public MainPage()
-	{
-		InitializeComponent();
-	}
-
-	private void OnCounterClicked(object sender, EventArgs e)
-	{
-
-	}
-}
-
